+++ GameEngine2.vcxproj
@@ -167,7 +167,7 @@
     <ClCompile Include="engine\Controller\GaemEngine.cpp" />
     <ClCompile Include="engine\Controller\InputManager.cpp" />
     <ClCompile Include="engine\Controller\Lua\LuaManger.cpp" />
-    <ClCompile Include="Engine\Controller\Physics.cpp" />
+    <ClCompile Include="Engine\Controller\Physics\Physics.cpp" />
     <ClCompile Include="Engine\Controller\ResourceManager.cpp" />
     <ClCompile Include="Engine\Controller\Serialization\SceneLoader.cpp" />
     <ClCompile Include="engine\Model\GameObject.cpp" />
@@ -191,7 +191,7 @@
     <ClCompile Include="Engine\View\Renderer.cpp" />
     <ClCompile Include="glad.c" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="engine\Model\Rigidbody.cpp" />
+    <ClCompile Include="Engine\Controller\Physics\Rigidbody.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="dependencies\include\imgui\imconfig.h" />
@@ -223,7 +223,7 @@
     <ClInclude Include="engine\Controller\InputManager.h" />
     <ClInclude Include="engine\Controller\Lua\LuaManager.h" />
     <ClInclude Include="Engine\Controller\md2Reader.h" />
-    <ClInclude Include="Engine\Controller\Physics.h" />
+    <ClInclude Include="Engine\Controller\Physics\Physics.h" />
     <ClInclude Include="Engine\Controller\ResourceManager.h" />
     <ClInclude Include="Engine\Controller\Serialization\SceneLoader.h" />
     <ClInclude Include="Engine\Controller\Timer.h" />
@@ -250,7 +250,7 @@
     <ClInclude Include="Engine\Model\Terrain.h" />
     <ClInclude Include="Engine\View\GUIRenderer.h" />
     <ClInclude Include="Engine\View\Renderer.h" />
-    <ClInclude Include="engine\Model\Rigidbody.h" />
+    <ClInclude Include="Engine\Controller\Physics\Rigidbody.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="dependencies\include\imgui\new 1.txt.bak" />
