--- data.json
@@ -1,5 +1,0 @@
-{
-"name": "Joe",
-"grade": "A",
-"words": ["word1", "word2"]
-}
