+++ engine/Controller/GaemEngine.cpp
@@ -71,9 +71,8 @@ GameEngine::GameEngine()
 	}
 
 	SceneLoader loader;
-
 	loader.SaveScene(&scene,"scene.json");
-	scene = *(loader.LoadScene("scene.json"));
+	//scene = *(loader.LoadScene("scene.json"));
 
 }
