+++ scene.json
@@ -121,9 +121,9 @@
 			"name" : "arcade1",
 			"position" : 
 			[
-				-298.0,
-				-32.680011749267578,
-				-1133.0
+				-444.0,
+				17.900001525878906,
+				359.0
 			],
 			"rigidbody" : 
 			{
