--- test.json
@@ -1,7 +1,0 @@
-{
-	"action" : "run",
-	"data" : 
-	{
-		"number" : 1
-	}
-}
